requirements.txt
mlflow==2.14.2
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
matplotlib==3.8.4
pyyaml==6.0.1
boto3==1.34.131
dvc==3.50.2
dvc-s3==3.2.0
fastapi==0.115.0
uvicorn==0.30.5
joblib==1.4.2